requirements.txt
ollama==0.6.1
flask==3.1.3
python-dotenv==1.2.2
pytest==9.0.2
